Add some effects with container
